requirements.txt
streamlit
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-qdrant==0.1.3
qdrant-client==1.11.1
python-dotenv
pypdf